Backend-hussain/requirements.txt
alembic==1.13.1
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.121.0
Flask==3.0.0
Flask-Cors==4.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httplib2==0.31.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
narwhals==1.34.1
numpy==2.2.4
oauthlib==3.3.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.4
psycopg[binary]
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.8.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
referencing==0.36.2
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.20
starlette==0.49.3
streamlit==1.44.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.38.0
watchdog==6.0.0
Werkzeug==3.0.1
